simplified file readers as a preparation to issue #7
removed the smart pointer from block_producer_wrapper, because this would lead to destruction of part of the NAPL graph when temporaries would be destroyed in statements such as:

new_sound = some_function( some_other_function( source));

(where some_other_function would return a block_producer_wrapper that would be destroyed before the end of the expression.

This fixed issue #12.

git-svn-id: svn://thuis/trunk/Sounds@120 3a7aa2ed-2c50-d14e-a1b9-f07bebb4988c