Change producer/consumer call sequence drastically
Calls to a producer RequestBlock() would trigger
one or more calls on the consumer's RecieveBlock().  This would lead to
a comples call stack that would go up and down in a graph of connected
producers and consumers.

The protocal is changed so that a producer will return a block of data
from the RequestBlock() call.  The returned block might not have all the
requested samples. It is up to the consumer to repeat calls to
RequestBlock() until it has all the samples it needs.

Also, it is no longer allowed for components to change the contents of a
block.